+++ scripts/remove-obsolete-backup-files.sh
@@ -10,6 +10,23 @@ then
     | head -n -$ARANGO_BACKUP_FILES_COUNT \
     | awk -v backup_path="$backups_dir" '{print "s3://"backup_path"/"$4}')
 
+  if [[ $? -ne 0 ]]
+  then
+    echo "Failed get list of backups"
+    exit 1
+  fi
+
   echo "Removing obsolete backup files"
   echo $files_to_remove | xargs -n1 aws s3 rm
+
+  if [[ $? -ne 0 ]]
+  then
+    echo "Failed remove obsolete backup files"
+    exit 1
+  fi
+
+  echo Obsolete backups are cleaned up
+
+else
+  echo Skip obsolete backups removing
 fi
